+++ R/cnd-entrace.R
@@ -28,7 +28,8 @@
 #' errors and warnings to be promoted to rlang conditions that include
 #' a backtrace.
 #'
-#' This is useful in conjunction with [`rlang_backtrace_on_error`].
+#' This is useful in conjunction with [`rlang_backtrace_on_error`] and
+#' [`rlang_backtrace_on_warning`].
 #'
 #' @section Under the hood:
 #' On R 4.0 and newer, `global_entrace()` installs a global handler