+++ R/cnd-message.R
@@ -325,6 +325,22 @@ on_load({
   })
 })
 
+on_load({
+  s3_register("knitr::sew", "rlang_warning", function(x, options, ...) {
+    # Simulate interactive session to prevent full backtrace from
+    # being included in error message
+    local_interactive()
+
+    # Include backtrace footer option in the condition. Processed by
+    # `cnd_message()`.
+    x <- cnd_set_backtrace_on_error(x, peek_backtrace_on_warning_report())
+
+    # The `sew.error()` method calls `as.character()`, which dispatches
+    # to `cnd_message()`
+    NextMethod()
+  })
+})
+
 #' Format bullets for error messages
 #'
 #' @description