+++ tests/testthat/test-trace.R
@@ -469,18 +469,6 @@ test_that("caught error does not display backtrace in knitted files", {
   })
 })
 
-test_that("can call `global_entrace()` in knitted documents", {
-  skip_if_not_installed("knitr")
-  skip_if_not_installed("rmarkdown")
-  skip_if(!rmarkdown::pandoc_available())
-
-  entrace_lines <- render_md("test-entrace.Rmd", env = current_env())
-
-  expect_snapshot({
-    cat_line(entrace_lines)
-  })
-})
-
 test_that("empty backtraces are dealt with", {
   foo <- NULL